Modified NavMesh compilation to use size_t
Earlier versions used unsigned int.  Apple version used uintptr_t.
size_t should be the most portable.

+++ src/Menge/MengeCore/resources/NavMesh.cpp
@@ -265,12 +265,12 @@ namespace Menge {
 				//		unsigned int into a pointer slot, it'll probably be safe.  Needs to be
 				//		tested.
 
-				unsigned int eID = reinterpret_cast< uintptr_t >( node._edges[ e ] );
+				size_t eID = reinterpret_cast< size_t >( node._edges[ e ] );
 				assert( eID >= 0 && eID < _eCount && "Finalizing invalid edge id from node" );
 				node._edges[ e ] = &_edges[ eID ];
 			}
 			for ( size_t o = 0; o < node._obstCount; ++o ) {
-				unsigned int oID = reinterpret_cast< uintptr_t >( node._obstacles[ o ] );
+				size_t oID = reinterpret_cast< size_t >( node._obstacles[ o ] );
 				assert( oID >= 0 && oID < _obstCount &&
 						"Finalizing invalid obstacle id for node" );
 				node._obstacles[ o ] = &_obstacles[ oID ];
@@ -282,11 +282,11 @@ namespace Menge {
 		// All of the node indices in the edges need to be replaced with pointers
 		for ( size_t e = 0; e < _eCount; ++e ) {
 			NavMeshEdge & edge = _edges[ e ];
-			unsigned int nID = reinterpret_cast< uintptr_t >( edge._node0 );
+			size_t nID = reinterpret_cast< size_t >( edge._node0 );
 			assert( nID >= 0 && nID < _nCount && "Finalizing invalid node id from edge" );
 			edge._node0 = &_nodes[ nID ];
 			
-			nID = reinterpret_cast< uintptr_t >( edge._node1 );
+			nID = reinterpret_cast< size_t >( edge._node1 );
 			assert( nID >= 0 && nID < _nCount && "Finalizing invalid node id from edge" );
 			edge._node1 = &_nodes[ nID ];
 			// compute edge distance
@@ -309,13 +309,13 @@ namespace Menge {
 			while ( curr != NavMeshObstacle::NO_NEIGHBOR_OBST && !processed[ curr ] ) {
 				processed[ curr ] = true;
 				NavMeshObstacle & obst = _obstacles[ curr ];
-				unsigned int nID = reinterpret_cast<uintptr_t>( obst._node );
+				size_t nID = reinterpret_cast<size_t>( obst._node );
                 //error: cast from 'Menge::Agents::Obstacle*' to 'unsigned int' loses precision
 				// [-fpermissive]
 				assert( nID < _nCount && "Finalizing invalid node id from obstacle" );
 				obst._node = &_nodes[ nID ];
 
-				nID = reinterpret_cast< uintptr_t >( obst._nextObstacle );
+				nID = reinterpret_cast< size_t >( obst._nextObstacle );
 				assert( ( nID < _obstCount || nID == NavMeshObstacle::NO_NEIGHBOR_OBST ) &&
 						"Finalizing invalid obstacle index for next obstacle" );
 				if ( nID == NavMeshObstacle::NO_NEIGHBOR_OBST ) {
